adding logic for q6 200 test

+++ models/articles-model.js
@@ -2,7 +2,7 @@ const db = require('../db/connection');
 
 
 const fetchArticles = () => {
-    return db.query(`SELECT articles.article_id, articles.title, articles.topic, articles.author, articles.created_at, articles.votes, articles.article_img_url, COUNT(comments.article_id) AS "comment_count" FROM articles LEFT JOIN comments ON comments.article_id = articles.article_id GROUP BY articles.article_id;`).then(({rows: articles}) => {
+    return db.query(`SELECT articles.article_id, articles.title, articles.topic, articles.author, articles.created_at, articles.votes, articles.article_img_url, COUNT(comments.article_id)::INT AS "comment_count" FROM articles LEFT JOIN comments ON comments.article_id = articles.article_id GROUP BY articles.article_id;`).then(({rows: articles}) => {
         return articles;
     });
